TF-IDF code for extracting key words
This Python script performs an analysis of authoritative reports categorized as Physical Risk and Transition Risk. Utilizing a TF-IDF model, it identifies the 50 most significant words from the specified sections of these reports, which are detailed in Appendix A of the associated thesis. These keywords are important for guiding in the extraction of relevant newspaper articles later on in the study. The script ensures accuracy by focusing only on specific sections of the reports, thereby enhancing the quality of the data.