README.md
Welcome to the official repository for the thesis titled "Climate Risk Through the Looking Glass: An Exploratory Analysis of Whether Climate Risks Are Priced into the Swedish Stock Market Using News-Based Indices and Asset Pricing Models." 
This repository is a collection of the code used during the development and analysis stages of the thesis. The primary goal of sharing  this repository is to provide a resource for other researchers and scholars who are exploring similar themes in financial economics, particularly those focused on the intersection of climate change and asset pricing.